pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "driftcut"
dynamic = ["version"]
description = "Early-stop decision gating for LLM model migrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Riccardo Merenda" },
]
keywords = ["llm", "migration", "evaluation", "canary", "early-stop"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.40",
    "httpx>=0.27",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "pre-commit>=3.7",
]

[project.scripts]
driftcut = "driftcut.cli:app"

[project.urls]
Homepage = "https://github.com/riccardomerenda/driftcut"
Repository = "https://github.com/riccardomerenda/driftcut"
Documentation = "https://docs.driftcut.dev"
Issues = "https://github.com/riccardomerenda/driftcut/issues"

[tool.hatch.version]
path = "src/driftcut/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/driftcut"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/driftcut/cli.py" = ["B008"]  # typer.Option() in defaults is idiomatic

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = ["yaml", "yaml.*", "redis", "redis.*"]
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
